refactor: add evaluate method for stage given speed and target discharge pressure
Today variable speed compressor trains use single speed compressor train at minimum
speed for INDIVIDUAL_ASV_PRESSURE. This should allow for making a better method
for variable speed compressor trains
