+++ src/libecalc/domain/process/compressor/core/train/variable_speed_compressor_train_common_shaft_multiple_streams_and_pressures.py
@@ -18,6 +18,7 @@ from libecalc.domain.process.compressor.core.train.stage import CompressorTrainS
 from libecalc.domain.process.compressor.core.train.types import (
     FluidStreamObjectForMultipleStreams,
 )
+from libecalc.domain.process.compressor.core.train.utils.common import EPSILON
 from libecalc.domain.process.compressor.core.train.utils.numeric_methods import (
     find_root,
     maximize_x_given_boolean_condition_function,
@@ -33,8 +34,6 @@ from libecalc.domain.process.compressor.dto import (
 from libecalc.domain.process.core.results.compressor import (
     TargetPressureStatus,
 )
-
-EPSILON = 1e-5
 
 
 class VariableSpeedCompressorTrainCommonShaftMultipleStreamsAndPressures(
